Move Sync API's to dedicated package.
