+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/internal/SyncSessionImpl.kt
@@ -23,7 +23,7 @@ import io.realm.internal.platform.freeze
 import io.realm.internal.util.Validation
 import io.realm.mongodb.SyncErrorCode
 import io.realm.mongodb.SyncException
-import io.realm.mongodb.SyncSession
+import io.realm.mongodb.sync.SyncSession
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.withContext
