+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/sync/SyncConfiguration.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.realm.mongodb
+package io.realm.mongodb.sync
 
 import io.realm.Configuration
 import io.realm.LogConfiguration
@@ -27,6 +27,10 @@ import io.realm.internal.platform.createDefaultSystemLogger
 import io.realm.internal.platform.singleThreadDispatcher
 import io.realm.log.LogLevel
 import io.realm.log.RealmLogger
+import io.realm.mongodb.App
+import io.realm.mongodb.Credentials
+import io.realm.mongodb.SyncException
+import io.realm.mongodb.User
 import io.realm.mongodb.internal.SyncConfigurationImpl
 import io.realm.mongodb.internal.UserImpl
 import kotlin.reflect.KClass
