+++ test/sync/src/androidTest/kotlin/io/realm/test/mongodb/shared/SyncSessionTests.kt
@@ -21,10 +21,10 @@ import io.realm.entities.sync.ChildPk
 import io.realm.entities.sync.ParentPk
 import io.realm.internal.interop.RealmInterop
 import io.realm.internal.platform.runBlocking
-import io.realm.mongodb.SyncConfiguration
 import io.realm.mongodb.SyncException
-import io.realm.mongodb.SyncSession
 import io.realm.mongodb.User
+import io.realm.mongodb.sync.SyncConfiguration
+import io.realm.mongodb.sync.SyncSession
 import io.realm.mongodb.syncSession
 import io.realm.query
 import io.realm.test.mongodb.TestApp
